fix: move pyplot usage to allow for batch and interactive plotting
